#376 - added the CacheProviderInterface->check() method, and read methods of the CacheProviderInterface now throw ResourceNotFoundException instances on cache read misses

+++ Alpha/Util/Cache/CacheProviderFile.php
@@ -2,6 +2,7 @@
 
 namespace Alpha\Util\Cache;
 
+use Alpha\Exception\ResourceNotFoundException;
 use Alpha\Util\Logging\Logger;
 use Alpha\Util\Config\ConfigProvider;
 
@@ -98,7 +99,7 @@ class CacheProviderFile implements CacheProviderInterface
         if (file_exists($filepath)) {
             return unserialize(file_get_contents($filepath));
         } else {
-            return false;
+            throw new ResourceNotFoundException('Unable to get a cache value on the key ['.$key.']');
         }
     }
 
@@ -125,6 +126,21 @@ class CacheProviderFile implements CacheProviderInterface
         $config = ConfigProvider::getInstance();
         $filepath = $config->get('app.file.store.dir').'/cache/files/'.$this->appPrefix.'-'.$key;
 
-        unlink($filepath);
+        if (file_exists($filepath)) {
+            unlink($filepath);
+        } else {
+            throw new ResourceNotFoundException('Unable to delete a cache value on the key ['.$key.']');
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function check($key): bool
+    {
+        $config = ConfigProvider::getInstance();
+        $filepath = $config->get('app.file.store.dir').'/cache/files/'.$this->appPrefix.'-'.$key;
+
+        return file_exists($filepath);
     }
 }
